Refocus project.
The purpose of Ogma is just to persist an in-memory KV store across app restarts. Security is not a concern.

Also removes all third-party dependencies.

Serialization is done with JSON as a compromise, and Brotli is used for compression for better ratios on JSON data. Brotli can be tuned via options using the CompressionLevel enum.